feat: First version of the oresreverter bot.
The current version is a single-threaded program which reads the config
from wiki, then runs in an infinite loop, printing all page titles for
which we encountered a revision with a damaging score above the
threshhold. If also has a basic backoff algorithm allowing it to
slow-down during slow periods.

Signed-off-by: Strainu <github@strainu.ro>